This stated as a conversion of monitor*.[ch] to the new buffer API, but it got away from me...
Includes a change from u_int -> size_t in the key API that
touches quite a few things

+++ ssh/authfd.h
@@ -40,8 +40,8 @@ int	ssh_remove_all_identities(int sock, int version);
 int	ssh_decrypt_challenge(int sock, struct sshkey* key, BIGNUM *challenge,
 	    u_char session_id[16], u_char response[16]);
 int	ssh_agent_sign(int sock, struct sshkey *key,
-	    u_char **sigp, u_int *lenp,
-	    u_char *data, u_int datalen, u_int compat);
+	    u_char **sigp, size_t *lenp,
+	    u_char *data, size_t datalen, u_int compat);
 
 /* Messages for the authentication agent connection. */
 #define SSH_AGENTC_REQUEST_RSA_IDENTITIES	1
